Phi-3-vision/requirements.txt
torch>=2.0.0
transformers>=4.36.0
Pillow>=10.0.0
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=4.25.0